refactor(prototyper): add platform mod
